Added page to list contents of files folder
This parses and loads the contents of the 'src/files' directory and
makes the file details available as 'files' eleventy data which is then
use to create an index page for the files directory, at the /files/ uri.
This also adds the new files index page as a footer link for easy
access.