Fix inconsistent get-involved div sizing
This means the contacts box is always wide enough for social media text/logos to not spill onto the next line, when screen width is over 1300px. The difference in minimal in Chromium/WebKit (a few pixels wider/narrower depending on the screen), but makes the box wide enough on Firefox, where it was too narrow before (Spanish text/logos were spilling over). Flex-basis is removed as it wasn't doing anything on Chromium/WebKit, and removing it helps make it display more similarly/better on Firefox

See https://github.com/OpenWebAdvocacy/website/issues/275